refactor: update schemas
