+++ apps/app/src/integrations/trpc/routers/operations.ts
@@ -1,8 +1,7 @@
 import { z } from "zod/v4";
 import { publicProcedure, router } from "../init";
 import { operationsService } from "../services";
-import { ObjectTypeSchema } from "@repo/schemas";
-import { OperationConfigSchema } from "@repo/pipeline-engine/schemas";
+import { ObjectTypeSchema, OperationConfigSchema } from "@repo/schemas";
 
 export const operationsRouter = router({
   getMany: publicProcedure.query(() => operationsService.getAll()),
