+++ apps/app/src/pages/OperationDetailPage/types.ts
@@ -4,8 +4,8 @@ export type {
   OutputPort,
   ExecutorConfig,
   OperationConfig,
-} from "@repo/pipeline-engine/schemas";
-import type { PortKind } from "@repo/pipeline-engine/schemas";
+} from "@repo/schemas";
+import type { PortKind } from "@repo/schemas";
 
 export const KIND_LABEL: Record<PortKind, string> = {
   text: "text",
