+++ packages/db-schema/package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "@repo/pipeline-engine": "workspace:*",
+    "@repo/schemas": "workspace:*",
     "drizzle-orm": "^0.45.2"
   },
   "devDependencies": {
