+++ packages/db-schema/src/tables/pipelines_table.ts
@@ -22,5 +22,4 @@ export const pipelinesTable = pgTable("pipelines", {
   createdAt: timestamp("created_at").notNull().defaultNow(),
   updatedAt: timestamp("updated_at").notNull().defaultNow(),
 });
-
 export type PipelineRecord = typeof pipelinesTable.$inferSelect;
