+++ packages/pipeline-engine/src/nodes/OperationNode.test.ts
@@ -2,7 +2,8 @@ import { describe, it, expect, vi, beforeEach } from "vitest";
 import { okAsync, errAsync } from "neverthrow";
 import { executeOperationNode, processOperationNode } from "./OperationNode";
 import type { PipelineEngineDeps } from "../deps";
-import type { PipelineNode, NodeCtx, ExecutorConfig } from "../schemas";
+import type { ExecutorConfig } from "@repo/schemas";
+import type { PipelineNode, NodeCtx } from "../schemas";
 import type { OperationNodeContext, OperationInfo } from "./types";
 
 vi.mock("@repo/obs", () => ({
