+++ packages/pipeline-engine/src/nodes/types.ts
@@ -1,5 +1,6 @@
 import type { PipelineEngineDeps } from "../deps";
-import type { NodeCtx, PipelineNode, OperationConfigInput } from "../schemas";
+import type { OperationConfigInput } from "@repo/schemas";
+import type { NodeCtx, PipelineNode } from "../schemas";
 import type { PipelineRunError } from "../errors";
 
 export type NodeResult = { ok: true } | { ok: false; error: PipelineRunError | null };
