+++ packages/pipeline-engine/src/schemas/BuiltinNodeTypeSchema.ts
@@ -1,0 +1,13 @@
+import { z } from "zod/v4";
+
+export const BuiltinNodeTypeSchema = z.enum([
+  "compound",
+  "condition",
+  "code-file",
+  "folder",
+  "github-project",
+  "operation",
+  "output-project-path",
+  "output-local-path",
+]);
+export type BuiltinNodeType = z.infer<typeof BuiltinNodeTypeSchema>;
