+++ packages/pipeline-engine/src/schemas/RunPromptOptionsSchema.ts
@@ -1,6 +1,6 @@
 import { z } from "zod/v4";
 
-import { AgentBackendSchema } from "./ExecutorConfigSchema";
+import { AgentBackendSchema } from "@repo/schemas";
 
 export const RunPromptOptionsSchema = z.object({
   prompt: z.string(),
