+++ packages/pipeline-engine/src/schemas/index.ts
@@ -1,3 +1,4 @@
+export * from "./BuiltinNodeTypeSchema";
 export * from "./NodeTypeSchema";
 export * from "./MetaNodeTypeSchema";
 export * from "./resolveMetaType";
@@ -16,11 +17,8 @@ export * from "./PipelineNodeDataSchema";
 export * from "./PipelineEdgeDataSchema";
 export * from "./PipelineNodeSchema";
 export * from "./PipelineEdgeSchema";
-export * from "./PortKindSchema";
-export * from "./InputPortSchema";
-export * from "./OutputPortSchema";
-export * from "./ExecutorConfigSchema";
-export * from "./OperationConfigSchema";
+export * from "./PipelineNodeSchema";
+export * from "./PipelineEdgeSchema";
 export * from "./NodeDataSchema";
 export * from "./NodeCtxSchema";
 export * from "./RunPromptOptionsSchema";
