+++ packages/pipeline-engine/src/schemas/resolveMetaType.ts
@@ -1,5 +1,5 @@
 import type { MetaNodeType } from "./MetaNodeTypeSchema";
-import type { BuiltinNodeType } from "./NodeTypeSchema";
+import type { BuiltinNodeType } from "./BuiltinNodeTypeSchema";
 
 const BUILTIN_META_MAP: Record<BuiltinNodeType, MetaNodeType> = {
   "code-file": "object",
