+++ packages/schemas/src/AgentBackendSchema.ts
@@ -1,0 +1,4 @@
+import { z } from "zod/v4";
+
+export const AgentBackendSchema = z.enum(["local-claude", "codex"]);
+export type AgentBackend = z.infer<typeof AgentBackendSchema>;
