+++ packages/schemas/src/ExecutorConfigSchema.ts
@@ -1,16 +1,8 @@
 import { z } from "zod/v4";
-
-export const ExecutorTypeSchema = z.enum(["agent", "script", "rule-check"]);
-export type ExecutorType = z.infer<typeof ExecutorTypeSchema>;
-
-export const AgentModeSchema = z.enum(["skill", "prompt"]);
-export type AgentMode = z.infer<typeof AgentModeSchema>;
-
-export const AgentBackendSchema = z.enum(["local-claude", "codex"]);
-export type AgentBackend = z.infer<typeof AgentBackendSchema>;
-
-export const ScriptLanguageSchema = z.enum(["bash", "python", "javascript"]);
-export type ScriptLanguage = z.infer<typeof ScriptLanguageSchema>;
+import { ExecutorTypeSchema } from "./ExecutorTypeSchema";
+import { AgentModeSchema } from "./AgentModeSchema";
+import { AgentBackendSchema } from "./AgentBackendSchema";
+import { ScriptLanguageSchema } from "./ScriptLanguageSchema";
 
 export const ExecutorConfigSchema = z.object({
   type: ExecutorTypeSchema,
