+++ packages/schemas/src/JobTypeSchema.ts
@@ -1,8 +1,5 @@
 import { z } from "zod/v4";
 
-export const JobStatusSchema = z.enum(["queued", "running", "done", "failed", "cancelled"]);
-export type JobStatus = z.infer<typeof JobStatusSchema>;
-
 export const JobTypeSchema = z.enum([
   "pipeline_run",
   "code_analysis",
