+++ packages/schemas/src/ObjectTypeSchema.ts
@@ -1,5 +1,6 @@
 import { z } from "zod/v4";
-import { OBJECT_TYPES } from "@repo/db-schema";
+
+export const OBJECT_TYPES = ["file", "folder", "project"] as const;
 
 export const ObjectTypeSchema = z.enum(OBJECT_TYPES);
 export type ObjectType = z.infer<typeof ObjectTypeSchema>;
