+++ packages/schemas/src/SkillSchema.ts
@@ -10,5 +10,4 @@ export const SkillSchema = z.object({
   createdAt: z.coerce.date(),
   updatedAt: z.coerce.date(),
 });
-
 export type Skill = z.infer<typeof SkillSchema>;
