+++ packages/schemas/src/index.ts
@@ -3,9 +3,19 @@ export * from "./BestPracticeImportEntrySchema";
 export * from "./BestPracticeImportSchema";
 export * from "./ChecklistItemSchema";
 export * from "./ChecklistResultSchema";
+export * from "./AgentBackendSchema";
+export * from "./AgentModeSchema";
 export * from "./CodeSnippetSchema";
+export * from "./ExecutorConfigSchema";
+export * from "./ExecutorTypeSchema";
 export * from "./GithubProjectSchema";
-export * from "./JobSchema";
+export * from "./ScriptLanguageSchema";
+export * from "./InputPortSchema";
+export * from "./OperationConfigSchema";
+export * from "./OutputPortSchema";
+export * from "./PortKindSchema";
+export * from "./JobStatusSchema";
+export * from "./JobTypeSchema";
 export * from "./ObjectTypeSchema";
 export * from "./OperationSchema";
 export * from "./PipelineSchema";
